requirement.txt
xopen
scipy
numpy
tklearn
biopython

